Fixed USB issue with newer HW. This should also improve throuput.
The data interrupt was not enabled.
This bug was overlooked, since SOF interrupt is triggered each 1ms,
so the interrupt is will eventually process the data related events.
The only problem can occur when device is not ready to handle incoming
SETUP packet (device is not allowed to send NACK).
Issue seems to appear on newer PCs with faster USB communication.